runpod_build/app/src/modules/AnimeCut/requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
moviepy>=1.0.3
Pillow>=10.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
